♻️ refactor: refactor the selectors import
