+++ src/store/session/index.ts
@@ -1,4 +1,3 @@
 export * from './hooks';
-export * from './selectors';
 export type { SessionStore } from './store';
 export { useSessionStore } from './store';
